Implemented simple inventory system
There is now a script to manage the player's inventory, as well as a script for any item objects. Inventory is represented by an integer array holding the number of each type of item collected. In the "Present" scene, there are squares the player can collect, which will cause their inventory to update.