Implemented Shop system
A shop consists of UI Buttons that are each assigned an item ID and a price. Click a button to buy an item, and the change in inventory and money is stored in InfoBank, to be passed into the player when they leave the shop.

Press S to switch to and from the shop scene.